Time: 0 ms (100%), Space: 42.2 MB (79.33%) - LeetHub

+++ 3622-check-divisibility-by-digit-sum-and-product/3622-check-divisibility-by-digit-sum-and-product.java
@@ -4,7 +4,7 @@ class Solution {
 
         int num = n;
         int sum = 0;
-        long product = 1;
+        int product = 1;
         int d = 0;
         while(num>0){
             d = num%10;
